serverless: rename files to try and fix CI errors
